Restructuralization of the project into modules
Also some minor changes and unification in DAO classes
