+++ persistence/src/main/java/backend/dao/ActivityRecordDAO.java
@@ -1,0 +1,8 @@
+package backend.dao;
+
+/**
+ * Interface for Activity Record Data Access Object.
+ * @author TODO
+ */
+public interface ActivityRecordDAO {
+}
