+++ persistence/src/main/java/backend/dao/BurnedCaloriesDAO.java
@@ -1,0 +1,48 @@
+package backend.dao;
+
+import backend.entities.BurnedCalories;
+
+import java.util.List;
+import java.util.UUID;
+
+/**
+ * Interface for Burned Calories Data Access Object.
+ * @author pmikova
+ */
+//TODO: This should be further discussed whether we dont want to link it with activity object?
+public interface BurnedCaloriesDAO {
+
+    /**
+     * Creates new BurnedCalories record.
+     * @param calories BurnedCalories object to be added to the database
+     */
+    void create(BurnedCalories calories);
+
+    /**
+     * Get burned calories by its ID.
+     * @param id of the calories object
+     * @return BurnedCalories object with given id
+     */
+    BurnedCalories getById(UUID id);
+
+    /**
+     * Get a list of all BurnedCalories objects.
+     * @return list of BurnedCalories objects
+     */
+    List<BurnedCalories> getAll();
+
+    /**
+     * Update burnedCalories record.
+     * @param calories BurnedCalories to update
+     */
+    void update(BurnedCalories calories);
+
+    /**
+     * Deletes BurnedCalories record.
+     * @param calories BurnedCalories to delete
+     */
+    void delete(BurnedCalories calories);
+
+}
+
+
