+++ persistence/src/main/java/backend/entities/ActivityRecord.java
