+++ persistence/src/main/java/backend/entities/BurnedCalories.java
@@ -1,0 +1,8 @@
+package backend.entities;
+
+/**
+ * This class should represent the calories table.
+ * @author TODO
+ */
+public class BurnedCalories {
+}
