--- src/main/java/backend/dao/ActivityRecordDAO.java
@@ -1,4 +1,0 @@
-package backend.dao;
-
-public interface ActivityRecordDAO {
-}
