--- src/main/java/backend/entities/BurnedCalories.java
@@ -1,4 +1,0 @@
-package backend.entities;
-
-public class BurnedCalories {
-}
